skillfinesse_app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
frozenlist==1.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
pdf2image==1.17.0
pillow==11.3.0
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.31.0
SQLAlchemy==2.0.40
firebase-admin==7.0.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.2.3
yarl==1.20.0